tests/wpt/mozilla/tests/css/writing-mode_change_display.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>writing-mode test</title>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
</head>
<body>
  <div>
    <span style="writing-mode: vertical-rl; border: 20px solid green">Text</span>
  </div>

  <script>
    test(function() {
      assert_equals(getComputedStyle(document.querySelector("span")).display, "inline-block");
    }, "The span's `display` should change from `inline` to `inline-block`, because of its writing-mode.");
  </script>
</body>
</html>
